--- applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/Themes/Standard/Components/Brand/MainNavbarBrandViewComponent.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Volo.Abp.AspNetCore.Mvc;
-
-namespace Unity.AspNetCore.Mvc.UI.Theme.UX2.Themes.Standard.Components.Brand;
-
-public class MainNavbarBrandViewComponent : AbpViewComponent
-{
-    public virtual IViewComponentResult Invoke()
-    {
-        return View("~/Themes/Standard/Components/Brand/Default.cshtml");
-    }
-}
