--- applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/Themes/Standard/Components/MainNavbar/Default.cshtml
@@ -1,22 +1,0 @@
-﻿@using Unity.AspNetCore.Mvc.UI.Theme.UX2.Themes.Standard.Components.Brand;
-@using Unity.AspNetCore.Mvc.UI.Theme.UX2.Themes.Standard.Components.Menu;
-@using Unity.AspNetCore.Mvc.UI.Theme.UX2.Themes.Standard.Components.Toolbar;
-
-<nav class="navbar navbar-expand-md shadow-sm flex-column flex-md-row unity-navbar" id="main-navbar">
-    <div class="container-fluid">
-        @(await Component.InvokeAsync<MainNavbarBrandViewComponent>())
-        <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
-                data-bs-target="#main-navbar-collapse" aria-controls="main-navbar-collapse"
-                aria-expanded="false" aria-label="Toggle navigation">
-            <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse" id="main-navbar-collapse">
-            <ul class="navbar-nav mx-auto">
-                @(await Component.InvokeAsync<MainNavbarMenuViewComponent>())
-            </ul> 
-            <ul class="navbar-nav"> 
-                @(await Component.InvokeAsync<MainNavbarToolbarViewComponent>())
-            </ul>
-        </div>
-    </div>
-</nav>
