--- applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/Themes/Standard/Components/Menu/MainNavbarMenuViewComponent.cs
@@ -1,22 +1,0 @@
-﻿using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Volo.Abp.AspNetCore.Mvc;
-using Volo.Abp.UI.Navigation;
-
-namespace Unity.AspNetCore.Mvc.UI.Theme.UX2.Themes.Standard.Components.Menu;
-
-public class MainNavbarMenuViewComponent : AbpViewComponent
-{
-    protected IMenuManager MenuManager { get; }
-
-    public MainNavbarMenuViewComponent(IMenuManager menuManager)
-    {
-        MenuManager = menuManager;
-    }
-
-    public virtual async Task<IViewComponentResult> InvokeAsync()
-    {
-        var menu = await MenuManager.GetMainMenuAsync();
-        return View("~/themes/Standard/Components/Menu/Default.cshtml", menu);
-    }
-}
