--- applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/Themes/Standard/Components/PageAlerts/PageAlertsViewComponent.cs
@@ -1,20 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Volo.Abp.AspNetCore.Mvc;
-using Volo.Abp.AspNetCore.Mvc.UI.Alerts;
-
-namespace Unity.AspNetCore.Mvc.UI.Theme.UX2.Themes.Standard.Components.PageAlerts;
-
-public class PageAlertsViewComponent : AbpViewComponent
-{
-    protected IAlertManager AlertManager { get; }
-
-    public PageAlertsViewComponent(IAlertManager alertManager)
-    {
-        AlertManager = alertManager;
-    }
-
-    public IViewComponentResult Invoke(string name)
-    {
-        return View("~/themes/Standard/Components/PageAlerts/Default.cshtml", AlertManager.Alerts);
-    }
-}
