--- applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/Themes/Standard/Components/Toolbar/LanguageSwitch/Default.cshtml
@@ -1,19 +1,0 @@
-﻿@using System.Linq
-@using Microsoft.AspNetCore.Http.Extensions
-@using Unity.AspNetCore.Mvc.UI.Theme.UX2.Themes.Standard.Components.Toolbar.LanguageSwitch;
-@model LanguageSwitchViewComponentModel
-@if (Model.OtherLanguages.Any())
-{
-    <div class="dropdown">
-        <a class="nav-link dropdown-toggle" href="#" role="button" id="dropdownMenuLink" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-            @Model.CurrentLanguage.DisplayName
-        </a>
-
-        <div class="dropdown-menu dropdown-menu-end border-0 shadow-sm" aria-labelledby="dropdownMenuLink">
-            @foreach (var language in Model.OtherLanguages)
-            {
-                <a class="dropdown-item" href="~/Abp/Languages/Switch?culture=@(language.CultureName)&uiCulture=@(language.UiCultureName)&returnUrl=@(System.Net.WebUtility.UrlEncode(Context.Request.GetEncodedPathAndQuery()))">@language.DisplayName</a>
-            }
-        </div>
-    </div>
-}
