--- applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/Themes/Standard/Components/Toolbar/LanguageSwitch/LanguageSwitchViewComponentModel.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections.Generic;
-using Volo.Abp.Localization;
-
-namespace Unity.AspNetCore.Mvc.UI.Theme.UX2.Themes.Standard.Components.Toolbar.LanguageSwitch;
-
-public class LanguageSwitchViewComponentModel
-{
-    public LanguageInfo CurrentLanguage { get; set; }
-
-    public List<LanguageInfo> OtherLanguages { get; set; }
-}
