--- applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/Themes/Standard/Components/Toolbar/MainNavbarToolbarViewComponent.cs
@@ -1,22 +1,0 @@
-﻿using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Volo.Abp.AspNetCore.Mvc;
-using Volo.Abp.AspNetCore.Mvc.UI.Theme.Shared.Toolbars;
-
-namespace Unity.AspNetCore.Mvc.UI.Theme.UX2.Themes.Standard.Components.Toolbar;
-
-public class MainNavbarToolbarViewComponent : AbpViewComponent
-{
-    protected IToolbarManager ToolbarManager { get; }
-
-    public MainNavbarToolbarViewComponent(IToolbarManager toolbarManager)
-    {
-        ToolbarManager = toolbarManager;
-    }
-
-    public virtual async Task<IViewComponentResult> InvokeAsync()
-    {
-        var toolbar = await ToolbarManager.GetAsync(StandardToolbars.Main);
-        return View("~/Themes/Standard/Toolbar/Default.cshtml", toolbar);
-    }
-}
