--- applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/Themes/Standard/Components/Topbar/Default.cshtml
@@ -1,37 +1,0 @@
-﻿@using Unity.AspNetCore.Mvc.UI.Theme.UX2.Themes.Standard.Components.Brand
-@using Volo.Abp.Users;
-
-@inject ICurrentUser CurrentUser
-
-<nav class="navbar navbar-expand-md shadow-sm flex-column flex-md-row unity-navbar" id="main-navbar">
-    <div class="container-fluid unity-navbar-btns">
-        @(await Component.InvokeAsync<MainNavbarBrandViewComponent>())
-
-        <button type="button" class="btn btn-nav" data-url="/Dashboard" onclick="window.location.href = '/Dashboard';">Dashboard</button>
-
-        <button type="button" class="btn btn-nav" data-url="/GrantApplications" onclick="window.location.href = '/GrantApplications';">Applications</button>
-
-        <button type="button" class="btn btn-nav" data-url="/Applicant" onclick="window.location.href = '';">Applicant</button>
-    </div>
-    <div class="unity-container-fluid">
-        <div class="image-with-text">
-            <img src="/themes/unity/images/union.png" alt="notifications">
-            <span class="text-inside-image">5</span>
-        </div>
-        <div class="px-3 d-flex align-items-center">
-            <div class="unity-user-initials">
-                @Html.Raw(CurrentUser.FindClaimValue("Badge"))
-            </div>
-        </div>
-    </div>
-    <abp-dropdown id="user-dropdown">
-        <abp-dropdown-button class="btn-dropdown" text="Ministry of Silly Walks - Universal Grant P..." />
-        <abp-dropdown-menu class="user-dropdown-items">
-            <abp-dropdown-item href="#">My account</abp-dropdown-item>
-            <abp-dropdown-item href="#">Grant Program Settings</abp-dropdown-item>
-            <abp-dropdown-item href="#">Logout</abp-dropdown-item>
-
-        </abp-dropdown-menu>
-    </abp-dropdown>
-</nav>
-
