--- applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/Themes/Standard/Layouts/Empty.cshtml
@@ -1,79 +1,0 @@
-﻿@using Unity.AspNetCore.Mvc.UI.Theme.UX2.Bundling;
-@using Unity.AspNetCore.Mvc.UI.Theme.UX2.Themes.Standard.Components.PageAlerts;
-
-@using Volo.Abp.AspNetCore.Mvc.UI.Components.LayoutHook;
-@using Volo.Abp.AspNetCore.Mvc.UI.Layout;
-@using Volo.Abp.AspNetCore.Mvc.UI.Theming
-@using Volo.Abp.AspNetCore.Mvc.UI.Widgets.Components.WidgetScripts
-@using Volo.Abp.AspNetCore.Mvc.UI.Widgets.Components.WidgetStyles
-@using Volo.Abp.Localization
-@using Volo.Abp.Ui.Branding
-@using Volo.Abp.Ui.LayoutHooks
-
-@inject IBrandingProvider BrandingProvider
-@inject IPageLayout PageLayout
-@{
-    Layout = null;
-    var containerClass = ViewBag.FluidLayout == true ? "container-fluid" : "container"; //TODO: Better and type-safe options
-
-    var pageTitle = ViewBag.Title == null ? BrandingProvider.AppName : ViewBag.Title; //TODO: Discard to get from Title
-
-    if (PageLayout.Content.Title != null)
-    {
-        if (!string.IsNullOrWhiteSpace(pageTitle))
-        {
-            pageTitle = " | " + pageTitle;
-        }
-
-        pageTitle = PageLayout.Content.Title + pageTitle;
-    }
-
-    var rtl = CultureHelper.IsRtl ? "rtl" : string.Empty;
-}
-
-<!DOCTYPE html>
-
-<html lang="@CultureInfo.CurrentCulture.Name" dir="@rtl">
-<head>
-    @await Component.InvokeLayoutHookAsync(LayoutHooks.Head.First, StandardLayouts.Empty)
-
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-
-    <title>@pageTitle</title>
-    @if (ViewBag.Description != null)
-    {
-        <meta name="description" content="@ViewBag.Description" />
-    }
-    <abp-style-bundle name="@UnityThemeUX2Bundles.Styles.Global" />
-
-    @await Component.InvokeAsync(typeof(WidgetStylesViewComponent))
-
-    @await RenderSectionAsync("styles", false)
-
-    @await Component.InvokeLayoutHookAsync(LayoutHooks.Head.Last, StandardLayouts.Empty)
-</head>
-<body class="abp-empty-layout @rtl">
-    @await Component.InvokeLayoutHookAsync(LayoutHooks.Body.First, StandardLayouts.Empty)
-
-    <div class="@containerClass">
-        @(await Component.InvokeAsync<PageAlertsViewComponent>())
-        @await Component.InvokeLayoutHookAsync(LayoutHooks.PageContent.First, StandardLayouts.Empty)
-        @RenderBody()
-        @await Component.InvokeLayoutHookAsync(LayoutHooks.PageContent.Last, StandardLayouts.Empty)
-    </div>
-
-    <abp-script-bundle name="@UnityThemeUX2Bundles.Scripts.Global" />
-
-    <script src="~/Abp/ApplicationLocalizationScript?cultureName=@CultureInfo.CurrentUICulture.Name"></script>
-    <script src="~/Abp/ApplicationConfigurationScript"></script>
-    <script src="~/Abp/ServiceProxyScript"></script>
-
-    @await RenderSectionAsync("scripts", false)
-
-    @await Component.InvokeAsync(typeof(WidgetScriptsViewComponent))
-
-    @await Component.InvokeLayoutHookAsync(LayoutHooks.Body.Last, StandardLayouts.Empty)
-</body>
-</html>
