+++ applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/Themes/UX2/Components/PageAlerts/PageAlertsViewComponent.cs
@@ -1,0 +1,20 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using Volo.Abp.AspNetCore.Mvc;
+using Volo.Abp.AspNetCore.Mvc.UI.Alerts;
+
+namespace Unity.AspNetCore.Mvc.UI.Theme.UX2.Themes.UX2.Components.PageAlerts;
+
+public class PageAlertsViewComponent : AbpViewComponent
+{
+    protected IAlertManager AlertManager { get; }
+
+    public PageAlertsViewComponent(IAlertManager alertManager)
+    {
+        AlertManager = alertManager;
+    }
+
+    public IViewComponentResult Invoke(string name)
+    {
+        return View("~/themes/UX2/Components/PageAlerts/Default.cshtml", AlertManager.Alerts);
+    }
+}
