+++ applications/Unity.GrantManager/modules/Unity.Theme.UX2/src/Unity.Theme.UX2/Themes/UX2/Components/SubTopbar/Default.cshtml
@@ -1,0 +1,58 @@
+﻿@using Unity.AspNetCore.Mvc.UI.Theme.UX2.Themes.UX2.Components.Brand;
+@using Volo.Abp.MultiTenancy;
+@using Volo.Abp.Users;
+
+@inject ICurrentTenant CurrentTenant
+@inject ICurrentUser CurrentUser
+
+@{  
+    string pageTitle = CurrentTenant.Name;
+    if (ViewBag.PageTitle != null) {
+        pageTitle += $" / {ViewBag.PageTitle}";
+    }
+}
+
+<nav class="navbar navbar-expand-md shadow-sm flex-column flex-md-row unity-sub-navbar" id="sub-navbar">
+    <div class="container-fluid">        
+        <h6>
+            @if (CurrentUser.IsAuthenticated
+            && CurrentUser.FindClaims("tenant").Length > 1)
+            {
+                <a class="navbar-nav-link-btn" href="/GrantPrograms" title="Switch Grant Programs"><i class="fl fl-mini-contract-2-mirrored"></i></a>
+            }
+            @pageTitle
+        </h6>
+        <div class="search-input" style="display:none">
+            <div class="input-group input-group-with-icons">
+                <div class="input-group-prepend">
+                    <span class="input-group-text">
+                        <i class="fs-5 fl fl-search"></i>
+                    </span>
+                </div>
+                <input type="text" id="search-bar" class="form-control" placeholder="Search" aria-label="Search" aria-describedby="basic-addon1">
+                <div class="input-group-append">
+                    <span class="input-group-text">
+                        <i class="fl fl-chevron-down"></i>
+                    </span>
+                </div>
+            </div>
+        </div>
+        <div class="d-flex">
+            <div class="" style="display:none">
+                <button type="button" class="btn btn-icon">
+                    <i class="fs-5 fl fl-settings"></i>
+                </button>
+            </div>
+            <div class="mx-1" style="display:none">
+                <button type="button" class="btn btn-icon">
+                    <i class="fs-5 fl fl-unknown"></i>
+                </button>
+            </div>
+            <div class="px-2 d-flex align-items-center">
+                <div class="unity-user-initials">                    
+                    @Html.Raw(CurrentUser.FindClaimValue("Badge"))                 
+                </div>
+            </div>
+        </div>
+    </div>
+</nav>
